Convert mixins to coremods (see description for why)
BlockModelMixin was originally used as a way to inject our own
ItemOverrides into whatever block model was being created. This worked
well, but the issue is that injecting at HEAD into a method can cause
compatibility issues with other injectors (be it mixin or ASM) that
attempt to modify other parts of these methods.

In most scenarios, the most effective solution would be to use
MixinExtras `@ModifyReturnValue` or `@WrapMethod` to wrap the entire
value into our own overrides instances. However, we can't since Forge
doesn't include MixinExtras. On top of that, not all Forge versions that
NEBs supports include Mixin at all (in our specific case, this is 1.14.4
and lower).

So, I believed that it might as well be easier to write coremods for all
versions of NEBs so that they can all execute on the same intention of
wrapping the original item overrides, instead of replacing the original
and consuming the list of overrides that would've been used by whatever
normally would've consumed it. This improves compatibility significantly
with other mods that might also attempt to do funky stuff with item
models, and is due to the coremods explicitely injecting bytecode that
modifies the value immediately before `ARETURN`, much like how
`@ModifyReturnValue` would.